+++ tests/TestCase.php
@@ -6,6 +6,7 @@ use Illuminate\Foundation\Testing\TestCase as BaseTestCase;
 use Laravel\Sanctum\Sanctum;
 use Modules\Identity\Domain\Models\User;
 use Modules\Catalog\Infrastructure\Persistence\Seeders\CatalogPermissionsSeeder;
+use Modules\Media\Infrastructure\Persistence\Seeders\MediaPermissionsSeeder;
 use Modules\Identity\Infrastructure\Persistence\Seeders\RolesAndPermissionsSeeder;
 
 abstract class TestCase extends BaseTestCase
@@ -18,6 +19,11 @@ abstract class TestCase extends BaseTestCase
     protected function seedCatalogPermissions(): void
     {
         $this->seed(CatalogPermissionsSeeder::class);
+    }
+
+    protected function seedMediaPermissions(): void
+    {
+        $this->seed(MediaPermissionsSeeder::class);
     }
 
     protected function actingAsCustomer(?User $user = null): User